refactor : asset의 .svg 이미지 파일 경로 변경
