+++ src/components/common/header/Header.tsx
@@ -11,11 +11,11 @@ import loadingImg from '../../../assets/loadingImg.svg';
 import { useModal } from '../../../hooks/useModal';
 import Modal from '../modal/Modal';
 import { formatImgPath } from '../../../util/formatImgPath';
-import bell from '../../../assets/bell.svg';
-import Notification from './Notification/Notification';
-import bellLogined from '../../../assets/bellLogined.svg';
-import { useEffect } from 'react';
-import { testLiveAlarm } from '../../../api/alarm.api';
+// import bell from '../../../assets/bell.svg';
+// import Notification from './Notification/Notification';
+// import bellLogined from '../../../assets/bellLogined.svg';
+// import { useEffect } from 'react';
+// import { testLiveAlarm } from '../../../api/alarm.api';
 import { useMyProfileInfo } from '../../../hooks/user/useMyInfo';
 import { ROUTES } from '../../../constants/user/routes';
