+++ src/components/user/comment/commentComponent/CommentComponentLayout.tsx
@@ -1,13 +1,13 @@
 import * as S from './CommentComponentLayout.styled';
 import DropDown from '../../../common/dropDown/DropDown';
 import DropDownItem from '../DropDownItem';
-import dropdownButton from '../../../assets/dropdownButton.svg';
 import useComment from '../../../../hooks/user/CommentHooks/useComment';
 import ReplyComponent from '../replyComponent/ReplyComponent';
-import ArrowDown from '../../../assets/ArrowDown.svg';
-import ArrowUp from '../../../assets/ArrowUp.svg';
+import ArrowUp from '../../../../assets/ArrowUp.svg';
 import CommentComponent from './commentComponent/CommentComponent';
 import { CommentType } from '../../../../models/comment';
+import dropdownButton from '../../../../assets/dropdownButton.svg';
+import ArrowDown from '../../../../assets/ArrowDown.svg';
 
 interface CommentLayoutProps {
   projectId: number;
