+++ src/components/user/home/banner/Banner.tsx
@@ -1,5 +1,5 @@
 import * as S from './Banner.styled';
-import banner from '../../../assets/banner.svg';
+import banner from '../../../../assets/banner.svg';
 
 export default function Banner() {
   return (
