+++ src/components/user/home/projectCardLists/cardList/CardList.tsx
@@ -1,10 +1,10 @@
 import PositionButton from '../../../../common/positionButton/PositionButton';
 import * as S from './CardList.styled';
-import beginner from '../../../../assets/beginner.svg';
 import { EyeIcon } from '@heroicons/react/24/outline';
 import type { ProjectList } from '../../../../../models/mainProjectLists';
 import { formatDate } from '../../../../../util/formatDate';
 import Avatar from '../../../../common/avatar/Avatar';
+import beginner from '../../../../../assets/beginner.svg';
 
 interface CardListProps {
   list: ProjectList;
