+++ src/components/user/home/searchFiltering/filteringContents/FilteringContents.tsx
@@ -1,6 +1,5 @@
 import Filtering from './filtering/Filtering';
 import * as S from './FilteringContents.styled';
-import beginner from '../../../../assets/beginner.svg';
 import { ChevronDownIcon } from '@heroicons/react/24/outline';
 import SkillTagBox from '../../../../common/skillTagBox/SkillTagBox';
 import React, { useState } from 'react';
@@ -8,6 +7,7 @@ import { useSearchFilteringSkillTag } from '../../../../../hooks/user/useSearchF
 import { useOutsideClick } from '../../../../../hooks/user/useOutsideClick';
 import { useSaveSearchFiltering } from '../../../../../hooks/user/useSaveSearchFiltering';
 import { SEARCH_FILTERING_DEFAULT_VALUE } from '../../../../../constants/user/homeConstants';
+import beginner from '../../../../../assets/beginner.svg';
 
 export default function FilteringContents() {
   const { positionTagsData, methodTagsData } = useSearchFilteringSkillTag();
